% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_admin_module.R
\name{custom_admin_module_ui}
\alias{custom_admin_module_ui}
\title{The UI for the "Admin Panel" dashboard}
\usage{
custom_admin_module_ui()
}
\arguments{
\item{id}{the Shiny module id.}
}
\value{
the UI for the "Admin Panel"
}
\description{
The \code{shiny} module UI for the \code{polished} Admin Panel, accessible to Admin users.
}
